update delete function from Angular project
Add dialog popup to let user confirm if they want to delete the selected record.

+++ AngularFolder/EmpReviewer/src/app/people-info/people-info.component.ts
@@ -1,7 +1,9 @@
+import { DialogDeleteComponent } from './../dialog-delete/dialog-delete.component';
 import { Router } from '@angular/router';
 import { PeopleService } from '../services/people.service';
 import { Component, OnInit } from '@angular/core';
 import { Person } from '../models/person';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-people-info',
@@ -16,6 +18,7 @@ export class PeopleInfoComponent implements OnInit {
 
   constructor(
     private _service: PeopleService,
+    private _dialog: MatDialog,
     private _router: Router) { }
 
   ngOnInit() {
@@ -26,15 +29,31 @@ export class PeopleInfoComponent implements OnInit {
     this.displayedColumns = ['id','firstName','lastName','age','detail','delete'];
   }
 
-  deletePerson(id:number){
-    this._service.deletePersonInfo(id).subscribe(()=> 
-    {
-      this.fetchData();
-    });
-  }
-
   navigateToDetailPage(id:number){
     this._router.navigate(['/people/' + id]);
+  }
+
+  deletePerson(id:number){
+    let dialogConfig = new MatDialogConfig();
+    let isDeleteRecord: boolean;
+
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+        
+    let result = this._dialog.open(DialogDeleteComponent, dialogConfig);
+
+    result.afterClosed().subscribe(r =>{
+      isDeleteRecord = r;
+
+      if(isDeleteRecord){
+        this._service.deletePersonInfo(id).subscribe(() => {
+          this.fetchData();
+        });
+      }
+      else{
+        console.log("cancle");
+      }
+    });
   }
 
   fetchData() {
@@ -42,4 +61,5 @@ export class PeopleInfoComponent implements OnInit {
       this.people = data;
     });
   }
+
 }
